Add metadata, reverse folder priority
This commit makes MSF look in the local data/meterpreter folder before
looking at the gem folder. This is to aid in development and testing,
as copying files into the appropriate gem folder is painful. Instead,
devs can copy directly to data/meterpreter and things will get picked
up from there. If missing, the gem takes over.

Details of the range of commits were added as metadata to the gemspec.
The commit hash which represents the commit for meterpreter that was
built into the gem is included, and the one for the previous release
is also included. This is so that we can track the range of commits
that the version contains.

Previous commit hash is not included in the range, but the current
one is.